backend/requirements.txt
arabic-reshaper
asgiref
asn1crypto
certifi
cffi
charset-normalizer
cryptography
cssselect2
Django
django-jet-reboot
django-simple-history
djangorestframework
djangorestframework_simplejwt
drf-yasg
freetype-py
html5lib
idna
inflection
lxml
MarkupSafe
oscrypto
packaging
pillow
pycairo
pycparser
pyHanko
pyhanko-certvalidator
PyJWT
pypdf
python-bidi
python-dotenv
python-http-client
pytz
PyYAML
reportlab
requests
rlPyCairo
sendgrid
six
sqlparse
svglib
tinycss2
tzdata
tzlocal
uritemplate
uritools
urllib3
webencodings
Werkzeug
xhtml2pdf
